+++ screens/TabOneScreen.tsx
@@ -3,8 +3,6 @@ import { StyleSheet } from "react-native"
 import EditScreenInfo from "../components/EditScreenInfo"
 import { Text, View } from "../components/Themed"
 
-import Button from "../components/Button/Button"
-import ButtonCourse from "../components/Button/ButtonCourse"
 export default function TabOneScreen({
   navigation,
 }: RootTabScreenProps<"TabOne">) {
@@ -17,8 +15,6 @@ export default function TabOneScreen({
         darkColor="rgba(255,255,255,0.1)"
       />
 
-      <Button title="Testing" />
-      <ButtonCourse title="CALCULO DIFERENCIAL" code="CC1035 - B" time="14:00 - 17:00" backgroundColor="#8572ff" color="#ffffff" />
       <EditScreenInfo path="/screens/TabOneScreen.tsx" />
     </View>
   )
